requirements.txt
# Beverly Knits AI Raw Material Planner - Requirements
# Note: For Python 3.13+, use requirements-python313.txt instead
# This file includes TensorFlow which requires Python <=3.12

# Core dependencies
pandas>=1.5.0
numpy>=1.23.0
asyncio-mqtt>=0.16.0

# ML/AI dependencies
scikit-learn>=1.3.0
tensorflow>=2.13.0  # Requires Python <=3.12
xgboost>=1.7.0
lightgbm>=4.0.0
statsmodels>=0.14.0
prophet>=1.1.0
scipy>=1.11.0

# Deep learning (optional, requires Python <=3.12)
torch>=2.0.0
torchvision>=0.15.0

# Data processing
openpyxl>=3.1.0
xlsxwriter>=3.1.0
pyarrow>=12.0.0

# API and async
aiohttp>=3.8.0
aiofiles>=23.0.0
httpx>=0.24.0
aiohttp==3.11.11  # For async HTTP requests to zen-mcp-server

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
pylint>=2.17.0

# Logging and monitoring
loguru>=0.7.0
python-json-logger>=2.0.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0

# Utilities
tqdm>=4.65.0
click>=8.1.0
rich>=13.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Additional ML utilities
joblib>=1.3.0
imbalanced-learn>=0.11.0
optuna>=3.2.0

# Data validation
great-expectations>=0.17.0
pandera>=0.16.0

# zen-mcp-server (if available)
# zen-mcp-server>=0.1.0